measurediff/pyproject.toml
[project]
name = "measurediff"
version = "0.1.0"
description = "Collect and compare Databricks Unity Catalog metric view measure definitions with full column lineage"
authors = [{ name = "Brook Patten", email = "brook.patten@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "databricks-connect==18.0.0",
    "pyyaml>=6.0",
    "sqlglot>=25.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.scripts]
measurediff = "measurediff.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest>=8.2.2"]

[tool.pytest.ini_options]
markers = [
    "integration: slower tests that require a live Databricks workspace",
    "unit: fast tests using only local and mocked resources",
]
